ページネーション、バリーデーション、ソート機能など実装

+++ app/Http/Controllers/BoardController.php
@@ -5,14 +5,27 @@ namespace App\Http\Controllers;
 use App\Models\Board;
 use App\Models\BoardComment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Schema;
 
 class BoardController extends Controller
 {
     public function index(Request $request)
     {
-        $select = Board::all();
+        $board_col = Schema::getColumnListing('board');
+        $result = in_array($request->sort, $board_col);
+
+        $board = Board::orderby(isset($request->sort) ? $request->sort : 'board_id', isset($request->order) ? $request->order : 'desc')->paginate(5);
+
+        if($result){
+            $board = Board::orderby($request->sort, $request->order)->paginate(5);
+            
+        } else {
+            return redirect()->back();
+        }
         $data = array(
-            'select' => $select,
+            'board' => $board,
+            'sort' => $request->sort,
+            'order' => $request->order,
         );
         return view('board.board_view', $data);
     }
@@ -20,6 +33,13 @@ class BoardController extends Controller
     public function content(Request $request)
     {
         $select = Board::find($request->board_id);
+
+        // 削除済みの投稿をクリックした時
+        if(empty($select->board_id)){
+            return redirect()->back()->withErrors('エラーが発生しました。');
+        }
+
+        // １：n ジョイン
         $select->comments();
 
         $data = array(
@@ -50,21 +70,38 @@ class BoardController extends Controller
         $board->note = $request->note;
         $board->save();
 
-        return redirect()->route('board.index');
+        return redirect()->route('board.index')->with('flash_message', '投稿を完了しました。');
     }
     
     public function delete(Request $request)
     {
         $board = Board::find($request->board_id);
+        // 投稿がない場合
+        if(empty($board)){
+            return redirect()->back()->withErrors('エラーが発生しました。');
+        }
+        // 投稿したユーザーと接続しているユーザーと違う場合
+        if($board->user_id != auth()->user()->id){
+            return redirect()->back()->withErrors('エラーが発生しました。');
+        }
+        
         $board->delete();
-        return redirect()->route('board.index');
+        return redirect()->route('board.index')->with('flash_message', '削除しました。');
     }
 
     public function comment_delete(Request $request)
     {
         $comment = BoardComment::find($request->comment_id);
+        if(empty($comment)){
+            return redirect()->back()->withErrors('エラーが発生しました。');
+        }
+        if($comment->user_id != auth()->user()->id){
+            return redirect()->back()->withErrors('エラーが発生しました。');
+        }
+
         $comment->delete();
-        return redirect()->back();
+        return redirect()->back()->with('flash_message', '削除しました。');
+
     }
 
     public function comment_create(Request $request)
@@ -78,7 +115,7 @@ class BoardController extends Controller
         $comment->board_id = $request->board_id;
         $comment->note = $request->comment_note;
         $comment->save();
-        return redirect()->back();
+        return redirect()->back()->with('flash_message', '投稿を完了しました。');
     }
 
     public function comment_update_ajax(Request $request)
